TLD can now grab webcam frames using Matlab without Image Acquisition package.
This works well under matlab 2010a, OpenCV 2.0+ and Ubuntu 10.04. The only thing the user has to do is execute ./server_stream before running TLD and set the in run_TLD.m the flag for camera capture.

This code is very close to working with Octave. Octave can succesfully grab the frames, but has problemas with the mouse input to draw the rectagle over the first object.

This version should have the same behavior as before for people using the Image Acquisition package. For this keep executing as before.